integrity-checker: Reintroduce an interface for comparing write sets.
Apparently this is used, just outside of this repository. Sorry for
breaking it.

CHANGELOG_BEGIN
CHANGELOG_END
