+++ ledger/participant-state/kvutils/tools/BUILD.bazel
@@ -37,7 +37,6 @@ da_scala_library(
         "//ledger/participant-integration-api",
         "//ledger/participant-state",
         "//ledger/participant-state/kvutils",
-        "//ledger/participant-state/kvutils:daml_kvutils_proto_java",
         "//libs-scala/concurrent",
         "//libs-scala/contextualized-logging",
         "//libs-scala/resources",
@@ -63,8 +62,6 @@ da_scala_test(
         "//ledger/participant-integration-api",
         "//ledger/participant-state/kvutils",
         "//ledger/participant-state/kvutils/tools",
-        "//ledger/participant-state/protobuf:ledger_configuration_proto_java",
-        "@maven//:com_google_protobuf_protobuf_java",
         "@maven//:org_mockito_mockito_core",
     ],
 )
