+++ ledger/participant-state/kvutils/tools/src/main/scala/ledger/participant/state/kvutils/tools/integritycheck/CommitStrategySupport.scala
@@ -5,22 +5,9 @@ package com.daml.ledger.participant.state.kvutils.tools.integritycheck
 
 import akka.stream.Materializer
 import com.daml.ledger.participant.state.kvutils.export.{SubmissionInfo, WriteSet}
-import com.daml.ledger.participant.state.v1.ReadService
 import com.daml.ledger.validator.StateKeySerializationStrategy
 
 import scala.concurrent.Future
-
-/** A ReadService that streams back previously recorded state updates */
-trait ReplayingReadService extends ReadService {
-  def updateCount(): Long
-}
-
-/** Records state updates and creates corresponding ReplayingReadService instances */
-trait ReplayingReadServiceFactory {
-  def appendBlock(writeSet: WriteSet): Unit
-
-  def createReadService(implicit materializer: Materializer): ReplayingReadService
-}
 
 trait CommitStrategySupport[LogResult] {
   def stateKeySerializationStrategy: StateKeySerializationStrategy
@@ -30,4 +17,6 @@ trait CommitStrategySupport[LogResult] {
   )(implicit materializer: Materializer): Future[WriteSet]
 
   def newReadServiceFactory(): ReplayingReadServiceFactory
+
+  def writeSetComparison: WriteSetComparison
 }
